requirements.txt
requests
beautifulsoup4
pytrends
python-slugify
